# docs/swagger.yaml
basePath: /
definitions:
...
  httptransport.AdminMFAStatusRequest:
    properties:
      user_id:
        type: string
    required:
    - user_id
    type: object
  httptransport.AdminResetMFARequest:
    properties:
      reason:
        type: string
      user_id:
        type: string
      version:
        type: integer
    required:
    - reason
    - user_id
    - version
    type: object
  httptransport.AdminResetMFAResponseBody:
    properties:
      reset:
        type: boolean
      revoked_sessions:
        type: integer
      user_id:
        type: string
      version:
        type: integer
    type: object
  httptransport.BatchGetIdentitiesRequest:
    properties:
      user_ids:
        items:
          type: string
        type: array
    required:
    - user_ids
    type: object
  httptransport.BatchGetServiceAccountsRequest:
    properties:
      service_account_ids:
        items:
          type: string
        type: array
    required:
    - service_account_ids
    type: object
...
  httptransport.CreateServiceAccountRequest:
    properties:
      audiences:
        items:
          type: string
        type: array
      name:
        type: string
    required:
    - audiences
    - name
    type: object
  httptransport.CreateServiceAccountResponseBody:
    properties:
      account:
        $ref: '#/definitions/httptransport.ServiceAccountResponseBody'
      client_secret:
        type: string
    type: object
...
  httptransport.GetIdentityRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
  httptransport.GetServiceAccountRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  httptransport.IdentityBatchResponseBody:
    properties:
      items:
        items:
          $ref: '#/definitions/httptransport.IdentityResponseBody'
        type: array
    type: object
  httptransport.IdentityPageResponseBody:
    properties:
      items:
        items:
          $ref: '#/definitions/httptransport.IdentityResponseBody'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  httptransport.IdentityResponseBody:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      phone:
        type: string
      status:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      username:
        type: string
      version:
        type: integer
    type: object
  httptransport.IssuePasswordResetRequest:
    properties:
      reason:
        type: string
      user_id:
        type: string
    required:
    - reason
    - user_id
    type: object
  httptransport.IssuePasswordResetResponseBody:
    properties:
      expires_at:
        type: string
      reset_token:
        type: string
    type: object
  httptransport.ListIdentitiesRequest:
    properties:
      keyword:
        type: string
      page:
        type: integer
      page_size:
        type: integer
      status:
        type: string
    type: object
...
  httptransport.ListServiceAccountsRequest:
    properties:
      keyword:
        type: string
      page:
        type: integer
      page_size:
        type: integer
      status:
        type: string
    type: object
...
  httptransport.MFAStatusResponseBody:
    properties:
      available:
        type: boolean
      enabled:
        type: boolean
      enabled_at:
        type: string
      recovery_codes_remaining:
        type: integer
      status:
        type: string
      version:
        type: integer
    type: object
...
  httptransport.RegisterIdentityRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - display_name
    - email
    - password
    - username
    type: object
  httptransport.Response:
    properties:
      body: {}
      code:
        type: integer
      message:
        type: string
      request_id:
        type: string
    type: object
...
  httptransport.RotateServiceAccountSecretRequest:
    properties:
      id:
        type: string
      version:
        type: integer
    required:
    - id
    - version
    type: object
  httptransport.RotateServiceAccountSecretResponseBody:
    properties:
      client_secret:
        type: string
      version:
        type: integer
    type: object
  httptransport.ServiceAccountBatchResponseBody:
    properties:
      items:
        items:
          $ref: '#/definitions/httptransport.ServiceAccountResponseBody'
        type: array
    type: object
  httptransport.ServiceAccountPageResponseBody:
    properties:
      items:
        items:
          $ref: '#/definitions/httptransport.ServiceAccountResponseBody'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  httptransport.ServiceAccountResponseBody:
    properties:
      audiences:
        items:
          type: string
        type: array
      client_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      version:
        type: integer
    type: object
...
  httptransport.UpdateIdentityProfileRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      phone:
        type: string
      reason:
        type: string
      version:
        type: integer
    required:
    - display_name
    - email
    - id
    - reason
    - version
    type: object
  httptransport.UpdateIdentityStatusRequest:
    properties:
      id:
        type: string
      reason:
        type: string
      status:
        type: string
      version:
        type: integer
    required:
    - id
    - status
    - version
    type: object
...
info:
  contact: {}
  description: 'Production-oriented Go Web API scaffold. Application error codes:
    0 success; 10000-19999 common/input; 20000-29999 authentication/authorization;
    30000-39999 business; 50000-59999 infrastructure.'
  title: Go API Template
  version: "1.0"
paths:
...
  /api/v1/identities/batch-get:
    post:
      consumes:
      - application/json
      parameters:
      - description: User IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.BatchGetIdentitiesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.IdentityBatchResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Batch get up to 100 user identities
      tags:
      - identities
  /api/v1/identities/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.GetIdentityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.IdentityResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Get a user identity by ID
      tags:
      - identities
  /api/v1/identities/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.ListIdentitiesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.IdentityPageResponseBody'
              type: object
      security:
      - Bearer: []
      summary: List user identities
      tags:
      - identities
  /api/v1/identities/mfa/reset:
    post:
      consumes:
      - application/json
      parameters:
      - description: Target user, reason, and expected MFA version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.AdminResetMFARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.AdminResetMFAResponseBody'
              type: object
        "409":
          description: 'Code 30009: stale resource version'
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Reset a user's MFA and revoke all sessions
      tags:
      - identities
  /api/v1/identities/mfa/status:
    post:
      consumes:
      - application/json
      parameters:
      - description: Target user
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.AdminMFAStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.MFAStatusResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Get a user's MFA status for platform administration
      tags:
      - identities
  /api/v1/identities/password-reset/issue:
    post:
      consumes:
      - application/json
      parameters:
      - description: Target user and auditable reason
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.IssuePasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.IssuePasswordResetResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Issue a one-time password recovery token for another user
      tags:
      - identities
  /api/v1/identities/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Identity
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.RegisterIdentityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.IdentityResponseBody'
              type: object
      security:
      - Bearer: []
      - PSK: []
      summary: Register a user identity
      tags:
      - identities
  /api/v1/identities/update-profile:
    post:
      consumes:
      - application/json
      parameters:
      - description: Profile, audit reason, and version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.UpdateIdentityProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.IdentityResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Update user profile with optimistic locking
      tags:
      - identities
  /api/v1/identities/update-status:
    post:
      consumes:
      - application/json
      parameters:
      - description: Status and version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.UpdateIdentityStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.IdentityResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Update user status with optimistic locking
      tags:
      - identities
...
  /api/v1/service-accounts/batch-get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Service account IDs (maximum 100)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.BatchGetServiceAccountsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.ServiceAccountBatchResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Get a bounded set of service accounts by ID
      tags:
      - service-accounts
  /api/v1/service-accounts/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Service account
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.CreateServiceAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.CreateServiceAccountResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Create a service account and return its secret once
      tags:
      - service-accounts
  /api/v1/service-accounts/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Service account ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.GetServiceAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.ServiceAccountResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Get a service account by ID
      tags:
      - service-accounts
  /api/v1/service-accounts/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.ListServiceAccountsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.ServiceAccountPageResponseBody'
              type: object
      security:
      - Bearer: []
      summary: List service accounts
      tags:
      - service-accounts
  /api/v1/service-accounts/rotate-secret:
    post:
      consumes:
      - application/json
      parameters:
      - description: Service account and version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.RotateServiceAccountSecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.RotateServiceAccountSecretResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Rotate a service-account secret with optimistic locking and return
        it once
      tags:
      - service-accounts
  /api/v1/service-accounts/update-status:
    post:
      consumes:
      - application/json
      parameters:
      - description: Status and version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.UpdateServiceAccountStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.UpdateServiceAccountStatusResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Enable or disable a service account with optimistic locking
      tags:
      - service-accounts
  /api/v1/sessions/get:
    post:
      consumes:
      - application/json
      parameters:
      - description: Session ID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.GetSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.SessionResponseBody'
              type: object
      security:
      - Bearer: []
      summary: Get an administrative session by ID
      tags:
      - sessions
  /api/v1/sessions/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: Filters and pagination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.ListSessionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.SessionPageResponseBody'
              type: object
      security:
      - Bearer: []
      summary: List user sessions for security administration
      tags:
      - sessions
  /api/v1/sessions/revoke:
    post:
      consumes:
      - application/json
      parameters:
      - description: Session, reason, and expected version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/httptransport.RevokeSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.Response'
            - properties:
                body:
                  $ref: '#/definitions/httptransport.SessionResponseBody'
              type: object
        "409":
          description: 'Code 30009: stale resource version'
          schema:
            $ref: '#/definitions/httptransport.Response'
      security:
      - Bearer: []
      summary: Administratively revoke a session with optimistic locking
      tags:
      - sessions
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Enter "Bearer {JWT}".
    in: header
    name: Authorization
    type: apiKey
  PSK:
    description: Enter "PSK {shared-key}" for routes configured with PSK authentication.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"